Reorganize ElGamal. Add reencryypt to ElGamalCiphertext.
